[S] JUnit boundary cases

+++ src/Storage/TNotesStorage.java
@@ -1,10 +1,10 @@
 package Storage;
 
-
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.assertNotNull;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -14,10 +14,8 @@ import java.util.Map;
 
 import Object.TaskFile;
 
-
 public class TNotesStorage {
 
-	
 	private static final String FILE_TYPE_TXT_FILE = ".txt";
 	private static final String MASTER_FILE_NAME = "masterfile.txt";
 	private static final String MAPPING_FILE_NAME = "FileToFolderMapping.txt";
@@ -29,12 +27,13 @@ public class TNotesStorage {
 	 * , sort alphabetically, and search. The file saves after every command
 	 * that the user types.
 	 * 
-	 * To Add: FolderManager class, change master list to contain name + date to facilitate this
+	 * To Add: FolderManager class, change master list to contain name + date to
+	 * facilitate this
 	 * 
 	 */
-	
+
 	private static TNotesStorage instance;
-	
+
 	private FolderManager fManager;
 	private ArrayList<String> masterList;
 	private ArrayList<String> floatingList;
@@ -43,16 +42,15 @@ public class TNotesStorage {
 	private File floatingListFile;
 	private File overviewFolder;
 	private Map<String, String> masterNameDateMap;
-	
 
 	// Constructor
 
 	private TNotesStorage() {
 		try {
-			
-			fManager = FolderManager.getInstance();		
+
+			fManager = FolderManager.getInstance();
 			overviewFolder = fManager.createDirectory(OVERVIEW_FILES_FOLDER_NAME);
-			
+
 			setUpMasterFile();
 			setUpFloatingTaskFileList();
 			masterNameDateMap = new HashMap<String, String>();
@@ -62,74 +60,69 @@ public class TNotesStorage {
 		}
 	}
 
-	
-	
-	
 	public static TNotesStorage getInstance() {
-		if(instance == null) {
+		if (instance == null) {
 			instance = new TNotesStorage();
 		}
 		return instance;
 	}
-	
-	
-	
+
 	private void setUpMasterFile() throws IOException {
-		
+
 		masterFile = createAnOverviewFile(MASTER_FILE_NAME);
-		
+
 		masterList = readFromMasterFile();
 	}
 
 	private void setUpFloatingTaskFileList() throws IOException {
 		floatingListFile = createAnOverviewFile(FLOATING_LIST_FILE_NAME);
-		
+
 		floatingList = readFromFloatingListFile();
 	}
-	
+
 	private File createAnOverviewFile(String name) throws IOException {
 		return fManager.createFile(overviewFolder, name);
-		
+
 	}
-	
-	
+
 	public void setUpMap() {
 		try {
-			
+
 			mapFile = createAnOverviewFile(MAPPING_FILE_NAME);
-		
+
 			masterNameDateMap = readFromMapFile();
-		
+
 		} catch (IOException ioEx) {
 			ioEx.printStackTrace();
 			System.err.println("HashMap error");
 		}
 	}
 
-	
 	public Map<String, String> readFromMapFile() throws IOException {
 		return fManager.readFromMapFile(mapFile);
 	}
-	
+
 	public boolean addTask(TaskFile task) {
-		
+
+		if (task.getName().length() > 260) {
+			return false;
+		}
+
 		if (!masterList.contains(task.getName())) {
 			masterList.add(task.getName());
 			String monthFolder;
-			
-			if(task.getIsTask()) {
+
+			if (task.getIsTask()) {
 				monthFolder = "floating";
 				floatingList.add(task.getName());
 				writeTaskToFloatingListFile(task);
-				
-				
+
 			} else {
 				monthFolder = getTaskMonth(task);
 			}
-			
+
 			masterNameDateMap.put(task.getName(), monthFolder);
-			
-			
+
 			if (writeTaskToMasterFile(task) && writeToMapFile(masterNameDateMap)) {
 				if (createTaskFile(monthFolder, task)) {
 					return true;
@@ -146,67 +139,79 @@ public class TNotesStorage {
 		return false;
 	}
 
-
 	private String getTaskMonth(TaskFile task) {
 		SimpleDateFormat monthStringFormat = new SimpleDateFormat("MMMM");
 		Calendar taskDate = task.getStartCal();
 		assertNotNull(taskDate);
-		
+
 		String monthFolder = monthStringFormat.format(taskDate.getTime());
 		return monthFolder;
 	}
 
 	public TaskFile deleteTask(String task) {
-		
-		TaskFile deletedTask = getTaskFileByName(task.trim());
-		
-		File fileToDelete = getTaskFilePath(task.trim());
-		
-		if (fManager.deleteFile(fileToDelete)) {
-			masterList.remove(task.trim());
-			writeListToMasterFile();
-			masterNameDateMap.remove(task.trim());
-			writeToMapFile(masterNameDateMap);
-			
-			if(deletedTask.getIsTask()) {
-				floatingList.remove(task.trim());
-				writeListToFloatingListFile();
+
+		try {
+			if (task.isEmpty()) {
+				return null;
 			}
-			return deletedTask;
+
+			TaskFile deletedTask = getTaskFileByName(task.trim());
+
+			File fileToDelete = getTaskFilePath(task.trim());
+
+			if (fManager.deleteFile(fileToDelete)) {
+				masterList.remove(task.trim());
+				writeListToMasterFile();
+				masterNameDateMap.remove(task.trim());
+				writeToMapFile(masterNameDateMap);
+
+				if (deletedTask.getIsTask()) {
+					floatingList.remove(task.trim());
+					writeListToFloatingListFile();
+				}
+				return deletedTask;
+			}
+			return null;
+		} catch (FileNotFoundException ioEx) {
+			System.err.println("File not Found");
+			return null;
 		}
-		return null;
 	}
 
 	public TaskFile getTaskFileByName(String taskName) {
-		
-		if (!masterList.contains(taskName.trim())) {
+
+		try {
+			if (!masterList.contains(taskName.trim())) {
+				throw new FileNotFoundException();
+			}
+
+			File fileToBeFound = getTaskFilePath(taskName.trim());
+
+			TaskFile taskFile = readTaskFile(fileToBeFound);
+
+			taskFile.setUpTaskFile();
+
+			return taskFile;
+		} catch (FileNotFoundException fNotFound) {
+			System.out.println("file does not exist");
 			return null;
 		}
-		
-		File fileToBeFound = getTaskFilePath(taskName.trim());
-		
-		TaskFile taskFile = readTaskFile(fileToBeFound);
-		
-		taskFile.setUpTaskFile();
-		
-		return taskFile;
 
 	}
 
-
-
-
-	private File getTaskFilePath(String taskName) {
-		
+	private File getTaskFilePath(String taskName) throws FileNotFoundException {
 
 		String folderName = masterNameDateMap.get(taskName.trim());
-		
-		//System.out.println(folderName + "3.");
-		//Check if folderName = recurr / floating
+
+		if (folderName == null || folderName.isEmpty()) {
+			throw new FileNotFoundException();
+		}
+		// System.out.println(folderName + "3.");
+		// Check if folderName = recurr / floating
 		File folder = fManager.appendParentDirectory(folderName);
-		
-		File fileToBeFound = fManager.getPathToFile(folder, taskName.trim()+ FILE_TYPE_TXT_FILE);
-		
+
+		File fileToBeFound = fManager.getPathToFile(folder, taskName.trim() + FILE_TYPE_TXT_FILE);
+
 		return fileToBeFound;
 	}
 
@@ -217,7 +222,7 @@ public class TNotesStorage {
 	public boolean clearAnOverviewFile(File fileToClear) {
 		return fManager.clearAnOverviewFile(fileToClear);
 	}
-	
+
 	public boolean clearMasterFile() {
 		return clearAnOverviewFile(masterFile);
 	}
@@ -225,19 +230,20 @@ public class TNotesStorage {
 	public boolean clearFloatingListFile() {
 		return clearAnOverviewFile(floatingListFile);
 	}
+
 	public boolean writeListToFloatingListFile() {
-		if(clearFloatingListFile()) {
+		if (clearFloatingListFile()) {
 			return fManager.writeListToFile(floatingListFile, floatingList);
 		}
-		
+
 		return false;
 	}
-	
+
 	public boolean writeListToMasterFile() {
-		if(clearMasterFile()) {
+		if (clearMasterFile()) {
 			return fManager.writeListToFile(masterFile, masterList);
 		}
-		
+
 		return false;
 	}
 
@@ -246,34 +252,32 @@ public class TNotesStorage {
 		String taskName = task.getName();
 		return fManager.writeTaskNameToListFile(masterFile, taskName);
 	}
-	
+
 	private boolean writeTaskToFloatingListFile(TaskFile task) {
 		String taskName = task.getName();
 		return fManager.writeTaskNameToListFile(floatingListFile, taskName);
 	}
-	
+
 	public boolean createTaskFile(String directory, TaskFile task) {
 
 		try {
 			File folder = fManager.createDirectory(directory.trim());
-			
+
 			File newTask = fManager.createFile(folder, task.getName() + FILE_TYPE_TXT_FILE);
 
 			return fManager.writeToTaskFile(newTask, task);
-			
-			
+
 		} catch (IOException ioEx) {
+			ioEx.printStackTrace();
 			System.err.println("file could not be created");
 			return false;
 		}
 	}
 
-	
-
 	public ArrayList<String> readFromMasterFile() {
 		return fManager.readFromListFile(masterFile);
 	}
-	
+
 	public ArrayList<String> readFromFloatingListFile() {
 		return fManager.readFromListFile(floatingListFile);
 	}
@@ -281,13 +285,14 @@ public class TNotesStorage {
 	public boolean clearMapFile() {
 		return clearAnOverviewFile(mapFile);
 	}
+
 	public boolean writeToMapFile(Map<String, String> map) {
-		if(clearMapFile()) {
+		if (clearMapFile()) {
 			return fManager.writeToMapFile(mapFile, map);
 		}
 		return false;
 	}
-	
+
 	public boolean clearMasterDirectory() {
 		return fManager.clearMasterDirectory();
 	}
